refactor: module.js에 함수 분리
